requirements.txt
fastapi
uvicorn
streamlit
requests
google-generativeai
python-dotenv
playwright
chromadb